Move base API JSON objects to separate file
