+++ src/planning_center_backend/_json_schemas/groups.py
@@ -5,25 +5,8 @@ from typing import Optional
 
 import msgspec
 
+from .base import _ApiBase, _DataBase
 
-# region Base Objects
-
-class _ApiBase(msgspec.Struct):
-    data: dict
-    included: list
-    meta: dict
-    links: Optional[dict] = None
-
-
-class _DataBase(msgspec.Struct, forbid_unknown_fields=True):
-    type: str
-    id: str
-    attributes: dict
-    links: Optional[dict] = None
-    relationships: Optional[dict] = None
-
-
-# endregion
 
 # region Groups
